#1 Bugfix: Fix sending messages (#2)
#1 Bugfix: Fix sending messages (#2)

What changed?
=====================

This commit fixes some issues outlined in issue #1, with creating new conversations and sending messages to others. It also resolves an issue with too many calls to the database via component re-renders.

Why?
====================

This is to allow users to create new conversations as well as send messages in existing conversations without errors or unexpected issues with the app.