#2 Feature: Implement a read status for messages (#4)
#2 Feature: Implement a read status for messages (#4)

What changed?
===================
This commit implements a read status for users by updating the seenBy column in the messages table. It also counts the number of messages that are unread and leaves an icon of the user if a message has been read.

Why?
===================
This feature is to allow users to see whether their message has been read or not. It has features that allow a user to see how many unread messages there are in a conversation and lets users see whether the other user has read a message.